producer/requirements.txt
kafka-python==2.0.2
pandas==2.0.0
numpy==1.24.0